Update roles and privileges tests.
