+++ src/lib/components/admin/configuration/RoleForm.svelte
@@ -91,7 +91,7 @@
   </fieldset>
 
   {#if validationError}
-    <aside class="alert variant-ghost-error">
+    <aside data-testid="validation-error" class="alert variant-ghost-error">
       <div class="alert-message">
         <p>{validationError}</p>
       </div>
